Started implementing process creation in userland

+++ Kernel/include/scheduler.h
@@ -20,9 +20,9 @@ int8_t get_state();
 void* scheduler(void *stack_pointer);
 PCBT *update_quantum(void *stack_pointer);
 
-uint16_t create_process(char *name, Priority priority, char foreground, char *argv[], int argc, main_function rip);
+uint64_t create_process(char *name, Priority priority, char foreground, char *argv[], int argc, main_function rip);
 void list_processes_state();
-uint8_t kill_process(unsigned int pid);
+uint64_t kill_process(unsigned int pid);
 uint16_t unblock_process(unsigned int pid);
 uint16_t block_process(unsigned int pid);
 void update_priority(unsigned int pid, Priority new_priority);
