+++ Userland/SampleCodeModule/include/syscalls.h
@@ -24,13 +24,13 @@ extern void call_sys_sound(uint32_t nFrequence, uint32_t time);
 extern void call_sys_checkHeight(char *HeightPassed, int command);
 extern void call_sys_draw_int(int number);
 
-extern uint8_t call_sys_kill_process(unsigned int pid);
+extern uint64_t call_sys_kill_process(unsigned int pid);
 extern void call_sys_update_priority(unsigned int pid, unsigned int new_priority);
 extern uint16_t call_sys_block_process(unsigned int pid);
 extern uint16_t call_sys_unblock_process(unsigned int pid);
 extern void call_sys_yield();
 extern void call_sys_process_status(unsigned int pid);
-extern uint16_t call_sys_create_process(char *name, unsigned int priority, char foreground, char *argv[], int argc, void *rip);
+extern uint64_t call_sys_create_process(char *name, unsigned int priority, char foreground, char *argv[], int argc, void *rip);
 extern void call_sys_list_processes_state();
 
 extern int64_t call_sys_get_pid();
